All warnings either fixed or supressed (supressed for code that we didn't write).

+++ src/cc_omniview_constants.h
@@ -28,17 +28,17 @@
 static const float kViewPoint_x_1 = 0.0;
 static const float kViewPoint_y_1 = -60.0;
 static const float kViewPoint_z_1 = 20;
-static const float kViewAngle_1 = M_PI/2;
-static const float kViewAngle_z_1 = -M_PI/8;
+static const float kViewAngle_1 = (float)(M_PI/2);
+static const float kViewAngle_z_1 = (float)(-M_PI/8);
 
 static const float kViewPoint_x_2 = 0.0;
 static const float kViewPoint_y_2 = -16.0;
 static const float kViewPoint_z_2 = 2.5;
-static const float kViewAngle_2 = M_PI/2;
+static const float kViewAngle_2 = (float)(M_PI/2);
 static const float kViewAngle_z_2 = 0.0;
 
 static const float kViewPoint_x_3 = 60.0;
 static const float kViewPoint_y_3 = -55.0;
 static const float kViewPoint_z_3 = 20.0;
-static const float kViewAngle_3 = 11.0*M_PI/16.0;
-static const float kViewAngle_z_3 = -M_PI/8.0;
+static const float kViewAngle_3 = (float)(11.0*M_PI/16.0);
+static const float kViewAngle_z_3 = (float)(-M_PI/8.0);
